Optimize trie to use less memory
The trie now works pretty much like a tree+hash-table hybrid; it's now
sort of a hash table that can perform prefix matches.

Previously, each trie node had an array with 256 pointers to other nodes
(which, not only wastes a lot of memory, but also means lookups often caused
a lot of expensive cache misses).  Each character from the key was an index
for this array to directly locate the next node.

Now, each node has 8 pointers, and each character is MOD 8 hashed.
Collisions are avoided by using a simple linked list, even though collisions
tend to be very rare considering the intended use-case as a matcher for
lwan_url_map_t.  Considering this, string comparisons are only performed if
there are collisions.

On 64-bit, the previous approach would use ~103kb of RAM for 10 keys with
average length of 5 characters.  This new approach uses less than 5kb with
the same result.  It is still a lot of memory, but an improvement
nonetheless.