Fix a information disclosure bug in file serving module
The reason a HTTP_NOT_FOUND is yielded here instead of a HTTP_FORBIDDEN is
that yielding HTTP_FORBIDDEN might lead to unwanted information disclosure
with malicious requests.

For example, if the request:

    GET /../../../../../../../../../../etc/debian_version HTTP/1.0

Yields a different response from:

    GET /../../../../../../../../../../etc/something_else HTTP/1.0

Then an attacker might know that this system is probably Debian based.

So just return HTTP_NOT_FOUND here -- which isn't wrong anyway, since the
requested resource is outside the root directory.