Add freegeoip sample application
Instructions to download the database can be found in this Git repository:
	git://github.com/fiorix/freegeoip

Copy over (or link) the `db` and `static` directories to where `freegeoip`
will be executed.  Currently, only the JSON output is supported.  Opening
the web page in a browser will show the same web page found on
http://freegeoip.net.

The only difference is that this version does not require a memcache server,
as it uses lwan's own naive implementation of one (which is a hash table
without collision management).  This is fine for a demonstration but will
most certainly not work well on any production environments.

Also note that sqlite3 is now required to build lwan.