Reduce epoll_ctl() chatter by using level-triggered behavior
After making a quick benchmark, here are the results for 100k requests
with 2k connections:
       With patch: 2004 calls to epoll_ctl()
    Without patch: 5183 calls to epoll_ctl()

Strace output is a lot nicer with this patch, but I couldn't measure
any significant different in throughput as I thought it would.  Since
all the tests are passing I'm pushing this.

I'm not really happy with how the CONN_FLIP_FLAGS thing, but I'll
eventually find a better way to do this.

Closes #125.