Rewrite the configuration file parser
Use a proper lexer+parser this time.  This removes all of the
limitations that the older parser imposed, so that things like:

	section {
		some_section param
		other section param2
		yet another section { with keys = yes }
	}

Are possible.  Note the lack of "{\n}" after the declaration of
"some_section" and "other", and the declaration of "yet" all in the
same line.  The testrunner.conf file has been modified to showcase the
new possibilities.

This uses the same lexer/parser technique as the template parser: two
state machines, with ring buffers, lexer calling the parser.  One of
the good things about this is that the configuration file is opened
once: isolating a section won't open it again; this fixes a possible
race condition where a file could be modified while a section was being
isolated.  The disadvantage is that the file must be memory-mappable,
but this is an implementation detail.

The API has been slightly modified but it's still essentially the same
thing.