Ensure stack is aligned on x86-64 and i386
On i386 it was already aligned, but on a 16-byte boundary by ANDing the
pointer with 0xfffffff0.  Bitwise AND it with 0xfffffffc, so it's
aligned on a 4-byte boundary.  (Save up to 12 bytes per coroutine stack
in the worst case! Maybe it's time to write a press release.)

On x86-64, it was only aligned by luck.  So mask the 4 least significant
bits from the stack pointer and align it on a 16-byte boundary, always.
This way, this shouldn't surprise and SIMD-using code.